lightnvm: add support for logical to physical map
For single block accesses, we encode the address directly in the nvme rw
cmd structure. The address (phys_slba) is stored in Dword 14-15 as a
unsigned 64 bit integer.

If the value is zero (and its a lightnvm device) its an error and the
command haven't been properly mapped and is aborted.

If value is > 0, its a valid address and it is mapped by the nvme
driver. The slba and elba is replace with the phy_slba to make
sure the appropriate disk address is written.

Please note that we overwrite the behavior of dword 14 and 15 and
thereby its not possible to use end-to-end protection.