nvme: fixed partitioning code.
LNVM would crash if the disk was of such a size, that the number
of LBA sectors (through nsze) did not completely align with the
number of blocks (pages != channel_len * nchannels).
This has now been fixed - moreover, this prepares the code for
actual erasure commands as the LBA's of each channel are now a
perfect multiple of the number of blocks assigned to it.